#128 - ft: Method DELETE: force ONLY ROLE_ADMIN

+++ src/Controller/ContratController.php
@@ -115,12 +115,14 @@ class ContratController extends AbstractController
     {
         $data = $request->toArray();
         if (isset($data['force']) && $data['force'] === true) {
+            if (!$this->isGranted("ROLE_ADMIN")) {
+                return new JsonResponse(["error" => "Hanhanhaaaaan vous n'avez pas dit le mot magiiiiqueeuuuuuh"], JsonResponse::HTTP_FORBIDDEN);
+            }
             $entityManager->remove($contrat);
         } else {
             $contrat->setStatus("off");
             $entityManager->persist($contrat);
         }
-
         $entityManager->flush();
         $cache->invalidateTags(["contrat"]);
         return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
