+++ src/Controller/ContratTypeController.php
@@ -92,18 +92,16 @@ class ContratTypeController extends AbstractController
     {
         $data = $request->toArray();
         if (isset($data['force']) && $data['force'] === true) {
+            if (!$this->isGranted("ROLE_ADMIN")) {
+                return new JsonResponse(["error" => "Hanhanhaaaaan vous n'avez pas dit le mot magiiiiqueeuuuuuh"], JsonResponse::HTTP_FORBIDDEN);
+            }
             $entityManager->remove($contratType);
         } else {
-            $contratType
-                ->setStatus("off")
-            ;
+            $contratType->setStatus("off");
             $entityManager->persist($contratType);
         }
-
         $entityManager->flush();
-
         $cache->invalidateTags(["contratType"]);
-
         return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
     }
 }
