+++ src/Controller/ProductTypeController.php
@@ -95,19 +95,14 @@ class ProductTypeController extends AbstractController
     {
         $data = $request->toArray();
         if (isset($data['force']) && $data['force'] === true) {
+            if (!$this->isGranted("ROLE_ADMIN")) {
+                return new JsonResponse(["error" => "Hanhanhaaaaan vous n'avez pas dit le mot magiiiiqueeuuuuuh"], JsonResponse::HTTP_FORBIDDEN);
+            }
             $entityManager->remove($productType);
-
-
         } else {
-            $productType
-                ->setStatus("off")
-            ;
-
+            $productType->setStatus("off");
             $entityManager->persist($productType);
         }
-
-
-
         $entityManager->flush();
         $cache->invalidateTags(["productType"]);
         return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
